fix FVCalc

+++ FVCalc.java
@@ -8,7 +8,7 @@ public class FVCalc {
 		double futureValue = currentValue * Math.pow(1+rate/100,n);
 
 		// System.out.println((int) futureValue);
-		System.out.println("After " + n + " years, $" + currentValue + " saved at " + rate + "% will yield $" + (int) futureValue);
+		System.out.println("After " + n + " years, $" + (int) currentValue + " saved at " + rate + "% will yield $" + (int) futureValue);
 
 		//After 2 years, $100 saved at 10.0% will yield $121
 	}
